requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
anyio==4.9.0
    # via
    #   -r requirements/quality.txt
    #   gql
asgiref==3.8.1
    # via
    #   -r requirements/quality.txt
    #   django
astroid==3.3.9
    # via
    #   -r requirements/quality.txt
    #   pylint
    #   pylint-celery
backoff==2.2.1
    # via
    #   -r requirements/quality.txt
    #   gql
build==1.2.2.post1
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools
cachetools==5.5.2
    # via
    #   -r requirements/ci.txt
    #   tox
cffi==1.17.1
    # via
    #   -r requirements/quality.txt
    #   pynacl
chardet==5.2.0
    # via
    #   -r requirements/ci.txt
    #   diff-cover
    #   tox
click==8.1.8
    # via
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   click-log
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
    #   pip-tools
click-log==0.4.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
code-annotations==2.2.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
colorama==0.4.6
    # via
    #   -r requirements/ci.txt
    #   tox
coverage[toml]==7.8.0
    # via
    #   -r requirements/quality.txt
    #   pytest-cov
diff-cover==9.2.4
    # via -r requirements/dev.in
dill==0.3.9
    # via
    #   -r requirements/quality.txt
    #   pylint
distlib==0.3.9
    # via
    #   -r requirements/ci.txt
    #   virtualenv
django==4.2.20
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/quality.txt
    #   django-crum
    #   django-waffle
    #   edx-django-utils
    #   edx-i18n-tools
django-crum==0.7.9
    # via
    #   -r requirements/quality.txt
    #   edx-django-utils
django-waffle==4.2.0
    # via
    #   -r requirements/quality.txt
    #   edx-django-utils
edx-django-utils==7.2.0
    # via -r requirements/quality.txt
edx-i18n-tools==1.7.0
    # via -r requirements/dev.in
edx-lint==5.6.0
    # via -r requirements/quality.txt
filelock==3.18.0
    # via
    #   -r requirements/ci.txt
    #   tox
    #   virtualenv
gql==3.5.2
    # via -r requirements/quality.txt
graphql-core==3.2.4
    # via
    #   -r requirements/quality.txt
    #   gql
idna==3.10
    # via
    #   -r requirements/quality.txt
    #   anyio
    #   yarl
iniconfig==2.1.0
    # via
    #   -r requirements/quality.txt
    #   pytest
isort==6.0.1
    # via
    #   -r requirements/quality.txt
    #   pylint
jinja2==3.1.6
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   diff-cover
lxml[html-clean,html_clean]==5.3.2
    # via
    #   edx-i18n-tools
    #   lxml-html-clean
lxml-html-clean==0.4.1
    # via lxml
markupsafe==3.0.2
    # via
    #   -r requirements/quality.txt
    #   jinja2
mccabe==0.7.0
    # via
    #   -r requirements/quality.txt
    #   pylint
multidict==6.4.2
    # via
    #   -r requirements/quality.txt
    #   yarl
newrelic==10.8.1
    # via
    #   -r requirements/quality.txt
    #   edx-django-utils
openedx-atlas==0.6.2
    # via -r requirements/quality.txt
packaging==24.2
    # via
    #   -r requirements/ci.txt
    #   -r requirements/pip-tools.txt
    #   -r requirements/quality.txt
    #   build
    #   pyproject-api
    #   pytest
    #   tox
path==16.16.0
    # via edx-i18n-tools
pbr==6.1.1
    # via
    #   -r requirements/quality.txt
    #   stevedore
pip-tools==7.4.1
    # via -r requirements/pip-tools.txt
platformdirs==4.3.7
    # via
    #   -r requirements/ci.txt
    #   -r requirements/quality.txt
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   -r requirements/ci.txt
    #   -r requirements/quality.txt
    #   diff-cover
    #   pytest
    #   tox
polib==1.2.0
    # via edx-i18n-tools
propcache==0.3.1
    # via
    #   -r requirements/quality.txt
    #   yarl
psutil==7.0.0
    # via
    #   -r requirements/quality.txt
    #   edx-django-utils
pycodestyle==2.13.0
    # via -r requirements/quality.txt
pycparser==2.22
    # via
    #   -r requirements/quality.txt
    #   cffi
pydocstyle==6.3.0
    # via -r requirements/quality.txt
pygments==2.19.1
    # via diff-cover
pylint==3.3.6
    # via
    #   -r requirements/quality.txt
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-django==2.6.1
    # via
    #   -r requirements/quality.txt
    #   edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   -r requirements/quality.txt
    #   pylint-celery
    #   pylint-django
pynacl==1.5.0
    # via
    #   -r requirements/quality.txt
    #   edx-django-utils
pyproject-api==1.9.0
    # via
    #   -r requirements/ci.txt
    #   tox
pyproject-hooks==1.2.0
    # via
    #   -r requirements/pip-tools.txt
    #   build
    #   pip-tools
pytest==8.3.5
    # via
    #   -r requirements/quality.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==6.1.1
    # via -r requirements/quality.txt
pytest-django==4.11.1
    # via -r requirements/quality.txt
python-slugify==8.0.4
    # via
    #   -r requirements/quality.txt
    #   code-annotations
pyyaml==6.0.2
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   edx-i18n-tools
six==1.17.0
    # via
    #   -r requirements/quality.txt
    #   edx-lint
sniffio==1.3.1
    # via
    #   -r requirements/quality.txt
    #   anyio
snowballstemmer==2.2.0
    # via
    #   -r requirements/quality.txt
    #   pydocstyle
sqlparse==0.5.3
    # via
    #   -r requirements/quality.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/quality.txt
    #   code-annotations
    #   edx-django-utils
text-unidecode==1.3
    # via
    #   -r requirements/quality.txt
    #   python-slugify
tomlkit==0.13.2
    # via
    #   -r requirements/quality.txt
    #   pylint
tox==4.25.0
    # via -r requirements/ci.txt
typing-extensions==4.13.1
    # via
    #   -r requirements/quality.txt
    #   anyio
virtualenv==20.30.0
    # via
    #   -r requirements/ci.txt
    #   tox
wheel==0.45.1
    # via
    #   -r requirements/pip-tools.txt
    #   pip-tools
yarl==1.19.0
    # via
    #   -r requirements/quality.txt
    #   gql

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
